Moved show deal command to deals show
