+++ commands/show.go
@@ -9,7 +9,6 @@ import (
 	"github.com/ipfs/go-ipfs-cmdkit"
 	"github.com/ipfs/go-ipfs-cmds"
 
-	"github.com/filecoin-project/go-filecoin/protocol/storage/storagedeal"
 	"github.com/filecoin-project/go-filecoin/types"
 )
 
@@ -19,7 +18,6 @@ var showCmd = &cmds.Command{
 	},
 	Subcommands: map[string]*cmds.Command{
 		"block": showBlockCmd,
-		"deal":  showDealCmd,
 	},
 }
 
@@ -64,57 +62,6 @@ Nonce:  %s
 				wStr,
 				strconv.FormatUint(uint64(block.Height), 10),
 				strconv.FormatUint(uint64(block.Nonce), 10),
-			)
-			return err
-		}),
-	},
-}
-
-var showDealCmd = &cmds.Command{
-	Helptext: cmdkit.HelpText{
-		Tagline: "Show deal details for CID <cid>",
-	},
-	Arguments: []cmdkit.Argument{
-		cmdkit.StringArg("cid", true, false, "CID of deal to query"),
-	},
-	Run: func(req *cmds.Request, re cmds.ResponseEmitter, env cmds.Environment) error {
-		propcid, err := cid.Decode(req.Arguments[0])
-		if err != nil {
-			return err
-		}
-
-		deal, err := GetPorcelainAPI(env).DealGet(req.Context, propcid)
-		if err != nil {
-			return err
-		}
-
-		if err := re.Emit(deal); err != nil {
-			return err
-		}
-		return nil
-	},
-	Type: storagedeal.Deal{},
-	Encoders: cmds.EncoderMap{
-		cmds.Text: cmds.MakeTypedEncoder(func(req *cmds.Request, w io.Writer, deal *storagedeal.Deal) error {
-			emptyDeal := storagedeal.Deal{}
-			if *deal == emptyDeal {
-				return fmt.Errorf("deal not found: %s", req.Arguments[0])
-			}
-
-			_, err := fmt.Fprintf(w, `Deal details
-CID: %s
-State: %s
-Miner: %s
-Duration: %d blocks
-Size: %s bytes
-Total Price: %s FIL
-`,
-				deal.Response.ProposalCid,
-				deal.Response.State,
-				deal.Miner.String(),
-				deal.Proposal.Duration,
-				deal.Proposal.Size,
-				deal.Proposal.TotalPrice,
 			)
 			return err
 		}),
