+++ commands/show_test.go
@@ -1,27 +1,15 @@
 package commands_test
 
 import (
-	"context"
-	"crypto/rand"
 	"encoding/json"
-	"io"
-	"math/big"
 	"testing"
 
-	"github.com/ipfs/go-cid"
-	files "github.com/ipfs/go-ipfs-files"
-	"github.com/multiformats/go-multihash"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/filecoin-project/go-filecoin/fixtures"
-	"github.com/filecoin-project/go-filecoin/proofs"
-	"github.com/filecoin-project/go-filecoin/protocol/storage/storagedeal"
 	th "github.com/filecoin-project/go-filecoin/testhelpers"
 	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
-	"github.com/filecoin-project/go-filecoin/tools/fast"
-	"github.com/filecoin-project/go-filecoin/tools/fast/fastesting"
-	"github.com/filecoin-project/go-filecoin/tools/fast/series"
 	"github.com/filecoin-project/go-filecoin/types"
 )
 
@@ -66,70 +54,4 @@ func TestBlockDaemon(t *testing.T) {
 
 		requireSchemaConformance(t, []byte(blockGetLine), "filecoin_block")
 	})
-}
-
-func TestShowDeal(t *testing.T) {
-	tf.IntegrationTest(t)
-
-	fastenvOpts := fast.FilecoinOpts{}
-
-	ctx, env := fastesting.NewTestEnvironment(context.Background(), t, fastenvOpts)
-	defer func() {
-		require.NoError(t, env.Teardown(ctx))
-	}()
-
-	clientNode := env.GenesisMiner
-	require.NoError(t, clientNode.MiningStart(ctx))
-
-	minerNode := env.RequireNewNodeWithFunds(1000)
-
-	// Connect the clientNode and the minerNode
-	require.NoError(t, series.Connect(ctx, clientNode, minerNode))
-
-	// Create a minerNode
-	collateral := big.NewInt(500)           // FIL
-	price := big.NewFloat(0.000000001)      // price per byte/block
-	expiry := big.NewInt(24 * 60 * 60 / 30) // ~24 hours
-
-	ask, err := series.CreateStorageMinerWithAsk(ctx, minerNode, collateral, price, expiry)
-	require.NoError(t, err)
-
-	// Create some data that is the full sector size and make it autoseal asap
-
-	maxBytesi64 := int64(getMaxUserBytesPerStagedSector())
-	dataReader := io.LimitReader(rand.Reader, maxBytesi64)
-	_, deal, err := series.ImportAndStore(ctx, clientNode, ask, files.NewReaderFile(dataReader))
-	require.NoError(t, err)
-
-	t.Run("showDeal outputs correct information", func(t *testing.T) {
-		showDeal, err := clientNode.ShowDeal(ctx, deal.ProposalCid)
-		require.NoError(t, err)
-
-		assert.Equal(t, ask.Miner, showDeal.Miner)
-		assert.Equal(t, storagedeal.Accepted, showDeal.Response.State)
-
-		duri64 := int64(showDeal.Proposal.Duration)
-		foo := big.NewInt(duri64 * maxBytesi64)
-
-		totalPrice := ask.Price.MulBigInt(foo)
-
-		assert.Equal(t, totalPrice, showDeal.Proposal.TotalPrice)
-	})
-
-	t.Run("When deal does not exist says deal not found", func(t *testing.T) {
-		deal.ProposalCid = requireTestCID(t, []byte("anything"))
-		showDeal, err := clientNode.ShowDeal(ctx, deal.ProposalCid)
-		assert.Error(t, err, "Error: deal not found")
-		assert.Nil(t, showDeal)
-	})
-}
-
-func getMaxUserBytesPerStagedSector() uint64 {
-	return proofs.GetMaxUserBytesPerStagedSector(types.OneKiBSectorSize).Uint64()
-}
-
-func requireTestCID(t *testing.T, data []byte) cid.Cid {
-	hash, err := multihash.Sum(data, multihash.SHA2_256, -1)
-	require.NoError(t, err)
-	return cid.NewCidV1(cid.DagCBOR, hash)
 }
